Rename logging package
